+++ sdk/keyvault/keyvault-certificates/recordings/node/certificates_client__list_certificates_in_various_ways/recording_list_deleted_certificates_with_requestoptions_timeout.js
@@ -1,3 +1,5 @@
 let nock = require('nock');
 
+module.exports.hash = "2d606eeaf8acd5274f073a3d1291a9a9";
+
 module.exports.testInfo = {"uniqueName":{},"newDate":{}}