+++ sdk/keyvault/keyvault-certificates/recordings/node/certificates_client__lro__recoverdelete/recording_can_recover_a_deleted_certificate_with_requestoptions_timeout.js
@@ -1,3 +1,5 @@
 let nock = require('nock');
 
+module.exports.hash = "1d277512bb04dd9cdf7874563f150fcb";
+
 module.exports.testInfo = {"uniqueName":{},"newDate":{}}