+++ sdk/keyvault/keyvault-certificates/src/core/challengeBasedAuthenticationPolicy.ts
@@ -10,21 +10,49 @@ import { WebResource } from "@azure/core-http";
 import { AccessTokenCache, ExpiringAccessTokenCache } from "@azure/core-http";
 
 /**
+ * Representation of the Authentication Challenge
+ */
+export class AuthenticationChallenge {
+  constructor(public authorization: string, public scope: string) {
+  }
+
+  /**
+   * Checks that this AuthenticationChallenge is equal to another one given.
+   * @param other The other AuthenticationChallenge
+   */
+  public equalTo(other: AuthenticationChallenge | undefined) {
+    if (!other) {
+      return false;
+    }
+    return this.authorization === other.authorization && this.scope === other.scope;
+  }
+}
+
+/**
+ * Helps keep a copy of any previous authentication challenges,
+ * so that we can compare on any further request.
+ */
+export class AuthenticationChallengeCache {
+  public challenge?: AuthenticationChallenge;
+
+  public setCachedChallenge(challenge: AuthenticationChallenge) {
+    this.challenge = challenge;
+  }
+}
+
+/**
  * Creates a new ChallengeBasedAuthenticationPolicy factory.
  *
  * @param credential The TokenCredential implementation that can supply the challenge token.
  */
 export function challengeBasedAuthenticationPolicy(credential: TokenCredential): RequestPolicyFactory {
   const tokenCache: AccessTokenCache = new ExpiringAccessTokenCache();
+  const challengeCache = new AuthenticationChallengeCache();
   return {
     create: (nextPolicy: RequestPolicy, options: RequestPolicyOptions) => {
-      return new ChallengeBasedAuthenticationPolicy(nextPolicy, options, credential, tokenCache);
+      return new ChallengeBasedAuthenticationPolicy(nextPolicy, options, credential, tokenCache, challengeCache);
     }
   };
-}
-
-export class AuthenticationChallenge {
-  constructor(public scopes: string[] | string) { }
 }
 
 /**
@@ -35,7 +63,6 @@ export class AuthenticationChallenge {
  *
  */
 export class ChallengeBasedAuthenticationPolicy extends BaseRequestPolicy {
-  private challenge: AuthenticationChallenge | undefined = undefined;
 
   /**
    * Creates a new ChallengeBasedAuthenticationPolicy object.
@@ -49,34 +76,42 @@ export class ChallengeBasedAuthenticationPolicy extends BaseRequestPolicy {
     nextPolicy: RequestPolicy,
     options: RequestPolicyOptions,
     private credential: TokenCredential,
-    private tokenCache: AccessTokenCache
+    private tokenCache: AccessTokenCache,
+    private challengeCache: AuthenticationChallengeCache
   ) {
     super(nextPolicy, options);
   }
 
-  private parseWWWAuthenticate(www_authenticate: string): string {
+  private parseWWWAuthenticate(www_authenticate: string): {
+    authorization: string,
+    resource: string
+  } {
+    const returnValue = {
+      authorization: "",
+      resource: ""
+    };
     // Parses an authentication message like:
     // ```
     // Bearer authorization="some_authorization", resource="https://some.url"
     // ```
-    let authenticateArray = www_authenticate.split(" ");
-
-    // Remove the "Bearer" piece
-    delete authenticateArray[0];
+    let spaceSep = www_authenticate.split(" ");
 
     // Split the KV comma-separated list
-    let commaSep = authenticateArray.join().split(",");
-    for (let item of commaSep) {
-      // Split the key/value pairs
-      let kv = item.split("=");
-      if (kv[0].trim() == "resource") {
-        // Remove the quotations around the string
-        let resource = kv[1].trim().replace(/['"]+/g, '');
-
-        return resource;
+    for (const spaceItem of spaceSep) {
+      const commaSep = spaceItem.split(",");
+      for (const commaItem of commaSep) {
+        // Split the key/value pairs
+        const kv = commaItem.split("=");
+        const key = kv[0].trim();
+        const removeQuotes = (x: string): string => x.trim().replace(/['"]+/g, '');
+        if (key == "authorization" || key == "authorization_uri") {
+          returnValue.authorization = removeQuotes(kv[1]);
+        } else if (key == "resource" || key == "scope") {
+          returnValue.resource = removeQuotes(kv[1]);
+        }
       }
     }
-    return "";
+    return returnValue;
   }
 
   /**
@@ -93,9 +128,9 @@ export class ChallengeBasedAuthenticationPolicy extends BaseRequestPolicy {
       throw new Error("The resource address for authorization must use the 'https' protocol.");
     }
 
-    let originalBody = webResource.body;
+    const originalBody = webResource.body;
 
-    if (this.challenge == undefined) {
+    if (this.challengeCache.challenge == undefined) {
       // Use a blank to start the challenge
       webResource.body = "";
     } else {
@@ -103,7 +138,7 @@ export class ChallengeBasedAuthenticationPolicy extends BaseRequestPolicy {
       await this.authenticateRequest(webResource);
     }
 
-    let response = await this._nextPolicy.sendRequest(webResource);
+    const response = await this._nextPolicy.sendRequest(webResource);
 
     if (response.status == 401) {
       webResource.body = originalBody;
@@ -111,17 +146,21 @@ export class ChallengeBasedAuthenticationPolicy extends BaseRequestPolicy {
       let www_authenticate = response.headers.get("WWW-Authenticate");
 
       if (www_authenticate) {
-        let resource = this.parseWWWAuthenticate(www_authenticate);
-        let challenge = new AuthenticationChallenge(resource + "/.default")
+        // The challenge based authentication will contain both an authorization URI with a token,
+        // and the resource to which that token is valid against (also called the scope).
+        const { authorization, resource } = this.parseWWWAuthenticate(www_authenticate);
+        const challenge = new AuthenticationChallenge(authorization, resource + "/.default")
 
-        if (this.challenge != challenge) {
-          this.challenge = challenge;
+        if (!challenge.equalTo(this.challengeCache.challenge)) {
+          this.challengeCache.setCachedChallenge(challenge);
           this.tokenCache.setCachedToken(undefined);
 
           await this.authenticateRequest(webResource);
+          return this._nextPolicy.sendRequest(webResource);
         }
+        return response;
       }
-      return this._nextPolicy.sendRequest(webResource);
+      return response;
     } else {
       return response;
     }
@@ -130,7 +169,7 @@ export class ChallengeBasedAuthenticationPolicy extends BaseRequestPolicy {
   private async authenticateRequest(webResource: WebResource): Promise<void> {
     let accessToken = this.tokenCache.getCachedToken();
     if (accessToken === undefined) {
-      accessToken = (await this.credential.getToken(this.challenge!.scopes)) || undefined;
+      accessToken = (await this.credential.getToken(this.challengeCache.challenge!.scope)) || undefined;
       this.tokenCache.setCachedToken(accessToken);
     }