+++ sdk/keyvault/keyvault-secrets/test/recoverBackupRestore.test.ts
@@ -76,7 +76,8 @@ describe("Secret client - restore secrets and recover backups", () => {
     } catch (e) {
       error = e;
     }
-    assert.equal(error.message, `Secret not found: ${secretName}`);
+    assert.equal(error.code, "SecretNotFound");
+    assert.equal(error.statusCode, 404);
   });
 
   if (isNode && !isPlaybackMode()) {
@@ -128,7 +129,8 @@ describe("Secret client - restore secrets and recover backups", () => {
     } catch (e) {
       error = e;
     }
-    assert.equal(error.message, `Secret not found: ${secretName}`);
+    assert.equal(error.code, "SecretNotFound");
+    assert.equal(error.statusCode, 404);
   });
 
   if (isRecordMode() || isPlaybackMode()) {